requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==23.2.0
Babel==2.10.3
black==25.1.0
click==8.2.1
decorator==5.2.1
Django==5.2.3
django-mixins==0.0.10
django-phonenumber-field==8.1.0
executing==2.2.0
flake8==7.3.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.17.0
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==9.0.10
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
wcwidth==0.2.13
